+++ src/R/r-utils.c
@@ -23,7 +23,7 @@ Riproc_hash_numeric (SEXP x)
 {
     iproc_vector_view view = Riproc_vector_view_sexp(x);
     size_t hash = iproc_vector_hash(&view.vector);
-    int int_hash = hash % INT_MAX;
+    int int_hash = (int)(hash % INT_MAX);
     return ScalarInteger(int_hash);
 }
 
@@ -59,7 +59,7 @@ Riproc_sexp2ptr (SEXP     s,
 SEXP
 Riproc_vector_new_copy (iproc_vector *vector)
 {
-    int64_t n = iproc_vector_dim(vector);
+    int n = (int)iproc_vector_dim(vector);
     SEXP Rvector;
 
     PROTECT(Rvector = NEW_NUMERIC(n));
@@ -72,8 +72,8 @@ Riproc_vector_new_copy (iproc_vector *vector)
 SEXP
 Riproc_matrix_new_copy (iproc_matrix *matrix)
 {
-    int64_t nrow = iproc_matrix_nrow(matrix);
-    int64_t ncol = iproc_matrix_ncol(matrix);
+    int nrow = (int)iproc_matrix_nrow(matrix);
+    int ncol = (int)iproc_matrix_ncol(matrix);
     SEXP Rmatrix;
 
     PROTECT(Rmatrix = allocMatrix(REALSXP, nrow, ncol));
@@ -86,7 +86,7 @@ Riproc_matrix_new_copy (iproc_matrix *matrix)
 iproc_vector_view
 Riproc_vector_view_sexp (SEXP Rvector)
 {
-    int64_t n = GET_LENGTH(Rvector);
+    int n = GET_LENGTH(Rvector);
     double *data = NUMERIC_POINTER(Rvector);
     return iproc_vector_view_array(data, n);
 }
@@ -95,13 +95,13 @@ iproc_matrix_view
 Riproc_matrix_view_sexp (SEXP Rmatrix)
 {
     if (isMatrix(Rmatrix)) {
-        int64_t nrow = INTEGER(GET_DIM(Rmatrix))[0];
-        int64_t ncol = INTEGER(GET_DIM(Rmatrix))[1];
+        int nrow = INTEGER(GET_DIM(Rmatrix))[0];
+        int ncol = INTEGER(GET_DIM(Rmatrix))[1];
         double *data = NUMERIC_POINTER(Rmatrix);
         return iproc_matrix_view_array(data, nrow, ncol);
     } else {
-        int64_t nrow = GET_LENGTH(Rmatrix);
-        int64_t ncol = 1;
+        int nrow = GET_LENGTH(Rmatrix);
+        int ncol = 1;
         double *data = NUMERIC_POINTER(Rmatrix);
         return iproc_matrix_view_array(data, nrow, ncol);
     }
